Re-org docker cli code
