+++ packages/dappmanager/src/modules/docker/compose/index.ts
@@ -1,0 +1,1 @@
+export * from "./cli";
