+++ packages/dappmanager/src/modules/installer/postInstallClean.ts
@@ -3,9 +3,9 @@ import path from "path";
 import * as db from "../../db";
 import { InstallPackageDataPaths } from "../../types";
 import { Log } from "../../utils/logUi";
-import { dockerCleanOldImages } from "../docker/dockerCommands";
 import { logs } from "../../logs";
 import { isNotFoundError } from "../../utils/node";
+import { dockerCleanOldImages } from "../docker/cleanOldImages";
 
 /**
  * [Post install clean] After a successful install, clean backup files
